+++ src/main/java/br/com/biblioteca/BibliotecaApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package br.com.biblioteca;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
