+++ src/main/java/br/com/biblioteca/models/LivroModel.java
@@ -1,4 +1,4 @@
-package com.example.demo.models;
+package br.com.biblioteca.models;
 
 import jakarta.persistence.*;
