+++ src/test/java/br/com/biblioteca/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package br.com.biblioteca;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
